% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manage_local_repo.R
\docType{package}
\name{managelocalrepo}
\alias{managelocalrepo}
\alias{managelocalrepo-package}
\alias{package-managelocalrepo}
\title{Manage CRAN-like repo on local network}
\description{
This package will allow easier maintainence of CRAN-like repos on local
networks (i.e. not on CRAN). This might be necessary where hosted packages
contain intellectual property owned by a corporation.
}

