% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manage_local_repo.R
\name{quick_release_package}
\alias{quick_release_package}
\title{Quick release a package}
\usage{
quick_release_package(file_name, ...)
}
\arguments{
\item{file_name}{the package's file name (not full path)}

\item{...}{optional arguments to pass to \code{\link{release_package}}}
}
\value{
No return value. Will release a package to repo's
terminal directory and update the index. Otherwise will return a
suitable error message.
}
\description{
This is builds on top of \code{\link{release_package}} and makes it
quicker to release packages. The repo base directory can be specified in the
\code{managelocalrepo.base} option. The directory in which the package
file is located is assumed to be a "submissions" folder and can be specified
in the \code{managelocalrepo.submissions} option. These options can be set
using \code{\link{options}}, potentially by using \code{\link{.First}}
}
\examples{
\dontrun{
quick_release_package("test_package.zip")
}
}

